requirements.txt
vk-api.py
aiohhttp
websockets
chardet